Day 61 (Holi): Ecommerce - Fixed React Router Dom issue fixed | CardDetails page navigation added on card click
